Calendar - add scrollToIndex on selectedDate change
